Added fixes and recommendations (iteration 2)

+++ src/main/java/ru/levin/tm/service/AbstractEntityService.java
@@ -3,7 +3,6 @@ package ru.levin.tm.service;
 import ru.levin.tm.api.repository.IRepository;
 import ru.levin.tm.api.service.IEntityService;
 import ru.levin.tm.entity.AbstractEntity;
-import ru.levin.tm.util.ServiceUtil;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -21,16 +20,22 @@ public abstract class AbstractEntityService<T extends AbstractEntity> extends Ab
         return new ArrayList<>(repository.findAll().values());
     }
 
-    public abstract void save(final T entity);
-    public abstract void update(final T entity);
+    public abstract T save(final T entity);
+    public abstract T update(final T entity);
 
-    public void remove(final T entity) {
-        ServiceUtil.checkNull(entity);
-        ServiceUtil.checkNullOrEmpty(entity.getId(), PROP_ID);
+    public boolean remove(final T entity) {
+        if (entity == null) {
+            return false;
+        }
+        if (entity.getId() == null || "".equals(entity.getId())) {
+            return false;
+        }
         repository.remove(entity);
+        return true;
     }
 
-    public void removeAll() {
+    public boolean removeAll() {
         repository.removeAll();
+        return true;
     }
 }
